pyproject.toml
[project]
name = "take-minutes"
dynamic = ["version"]
description = "Extract structured knowledge from conversation transcripts"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Daniel Iser"}]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing",
]

dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/danieliser/take-minutes"
Repository = "https://github.com/danieliser/take-minutes"
Issues = "https://github.com/danieliser/take-minutes/issues"

[project.optional-dependencies]
search = [
    "numpy>=2.0",
    "faiss-cpu>=1.9",
    "sentence-transformers>=2.2",
]

[dependency-groups]
dev = ["pytest>=7.0"]

[project.scripts]
minutes = "minutes.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/minutes/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/minutes"]

[tool.pytest.ini_options]
testpaths = ["tests"]
